Now Works Without Command Line Arguments

+++ question.cpp
@@ -1,14 +1,12 @@
 #include "question.hpp"
 
-Question::Question(int new_first, int new_second, char op) {
+Question::Question(double new_first, double new_second, char op) {
     first = new_first;
     second = new_second;
     operation = op;
 }
 
 bool Question::check(int time, int max_time) {
-    // correct = (student_answer == first * second) && (time < 5000000);
-    // return correct;
     bool correct_answer;
     switch (operation) {
         case '+':
@@ -39,8 +37,6 @@ bool Question::check(int time, int max_time) {
 }
 
 void Question::getStudentAnswer() {
-    //t1 = std::thread([this](){std::cin >> student_answer; cv.notify_one();});
-    //std::this_thread::sleep_for(std::chrono::milliseconds(5000));
     std::string answer;
     std::getline(std::cin, answer);
     student_answer = answer.empty() ? -1 : std::stoi(answer);
